# [experimenta] Optimize value.reflector types that implement json.Marshal