Reduces ObjectToTyped latency by ~15% and allocation count by about ~20%.

Optimization:

```
go test ./vendor/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager -run=NONE -bench=BenchmarkConvertObjectToTyped -memprofile=mem.prof -cpuprofile=cpu.prof -benchtime=5s
 goos: linux
 goarch: amd64
 pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager
 BenchmarkConvertObjectToTyped/Pod/structured-12                    65568             90843 ns/op            8614 B/op        263 allocs/op
 BenchmarkConvertObjectToTyped/Node/structured-12                   34234            168952 ns/op           20955 B/op        593 allocs/op
 BenchmarkConvertObjectToTyped/Endpoints/structured-12               1521           3360417 ns/op          258148 B/op       7036 allocs/op
```

Baseline:

```
go test ./vendor/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager -run=NONE -bench=BenchmarkConvertObjectToTyped -memprofile=mem.prof -cpuprofile=cpu.prof -benchtime=5s
 goos: linux
 goarch: amd64
 pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager
 BenchmarkConvertObjectToTyped/Pod/structured-12                    49263            107830 ns/op            9654 B/op        280 allocs/op
 BenchmarkConvertObjectToTyped/Node/structured-12                   27330            212387 ns/op           27908 B/op        711 allocs/op
 BenchmarkConvertObjectToTyped/Endpoints/structured-12               1504           3940001 ns/op          258377 B/op       7039 allocs/op
```

TODO:

- [ ] The copy-on-write may perform poorly for the full set of types in k8s (the copies might get expensive). Can we either pre-populate the cache by scanning all native types, or do a more efficient copy-on-write approach?
- [ ] This requires that `UnstructuredStringConverter` be somehow provided by k8s, but we haven't worked out a good way to do that yet, right now we just have k8s code call Register and provide impls for time and quantity